Production DB support (postgres) (#103)
Defines a Storage interface and adds support for choosing between Bolt and Postgres storage types.

* BoltDB code has been mostly left as is, with modifications of adding unused context arguments to function signatures to make it conform to the interface, making functions that don't conform to the interface private (to discourage skipping the interface) and removing functions that are unused.
* Postgres support uses sqlc to generate query wrappers and goose for migrations.

A switch statement takes a URI for the storage provider, and returns either bolt or postgres based on the URI scheme. This creates a backwards-incompatible change requiring the configured Bolt path to be prefixed with bolt://. If that is an issue, adding a special case mapping of no scheme -> bolt would be pretty easy. I also renamed the config field from db_file to storage_uri.

Fixes #41